Added roles and permissions and laravel to typescript model generation

+++ app/Http/Controllers/PeertjesController.php
@@ -4,13 +4,15 @@ namespace App\Http\Controllers;
 
 use App\Models\Peertje;
 use App\Models\PeertjeLocation;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
 
 class PeertjesController extends Controller
 {
+    public function index() {
+        $this->authorize('viewAny', Peertje::class);
 
-    public function show() {
         $peertjes = Peertje::query()->whereHas('locations', function ($query) {
             $query->where('created_at', '>', now()->subDays(7));
         })->get();
@@ -20,7 +22,33 @@ class PeertjesController extends Controller
         ]);
     }
 
+    public function show(Peertje $peertje) {
+        $this->authorize('update', $peertje);
+
+
+        return Inertia::render('Peertjes/Show', [
+            'peertje' => $peertje,
+        ]);
+    }
+
+    public function update(Request $request, Peertje $peertje) {
+        $this->authorize('update', $peertje);
+
+        $validated = $request->validate([
+            'name' => 'required|string',
+            'api_id' => 'integer|between:0,7|unique:peertjes',
+        ]);
+
+        $peertje->update($validated);
+
+        return redirect()->back();
+    }
+
     public function list() {
+        if(!\Auth::user()->can('manage-peertjes')) {
+            abort(403, 'Je hebt niet de rechten om peertjes te bewerken.');
+        }
+
         $peertjes = Peertje::all();
 
         return Inertia::render('Peertjes/List', [
@@ -29,23 +57,29 @@ class PeertjesController extends Controller
     }
 
     public function create(Request $request) {
+        $this->authorize('create', Peertje::class);
+
         $validated = $request->validate([
             'name' => 'required|string',
+            'api_id' => 'integer|between:0,7|unique:peertjes',
         ]);
 
-        $peertje = new Peertje();
-        $peertje->name = $validated['name'];
+        $peertje = new Peertje($validated);
         $peertje->save();
 
-        return redirect()->route('peertjes.list');
+        return redirect()->back();
     }
 
     public function destroy(Request $request) {
+
         $validated = $request->validate([
             'id' => 'required|integer|exists:peertjes,id',
         ]);
 
         $peertje = Peertje::query()->findOrFail($validated['id']);
+
+        $this->authorize('delete', $peertje);
+
         $peertje->delete();
 
         return redirect()->route('peertjes.list');
